requirements.txt
streamlit
pandas
PyPDF2
influxdb-client
plotly
camelot-py[cv]
python-dotenv
